Build testfloat_gen, needed to run the tests
In order to run the standalone tests we need to be able to run
"testfloat_gen", part of the TestFloat-3 package.  Since this is
probably going to be something that's not commonly installed, I added
some Makefile targets that download and build "testfloat_gen" so it
can be run without messing around with any system dependencies.

Note that TestFloat-3 by default builds in 8086 mode, so there's a
patch here that makes it build in RISC-V mode instead.