Re-apply my "build testfloat_gen" patch
I'm not actually sure what happened here, but it appears one of my patches has
been lost.  The RISCV-specific patch is still there, and it's in "git log", but
it's not in the current Makefile.  "git blame" points to the patch that
contains the code I changed, which I wouldn't expect.

This was originally part of b8cfc323de5adf79b1c9fca8e24be713850e392d.