Stop patching berkeley-softfloat-3
Newer releases of berkeley-softfloat-3 allow users to set the
SPECIALIZE_TYPE variable without applying a patch.  This commit removes
the patch and sets the variable on the make commandline instead.

Note that there's no dependency checking for berkeley-softfloat-3 when
checked out locally, so you might have to "make distclean" to get the
new version.