Proof of concept for MulAdd unit with additional Mul-only output
The Mul output is ready many gate delays before the MulAdd output,
so may be useful in some situations.

Not sure whether this is the best way to factor the code, so I'm not
going to PR this yet, but am pushing it so it doesn't get lost.

Tested with testfloat_gen -level 2.